Add layout-id “73” for the ALC283
This commit adds the layout-id “73” for the ALC283/ALC3239, made specifically for the Lenovo ThinkCentre M73 Tiny.

This layout-id fixes having to decide between having a non-working microphone jack (for this, I used layout-id 88), or muffled headphone audio (for this, I used layout-id 66). This config features both of those features working fine, and through my testing, works as well (or in some cases, better) then how sound works on Windows for this machine. It still does have issues in my experience (only the left-channel records through the microphone jack), but I’m assuming that is a hardware issue, because it also happens in Windows.

Signed-off-by: UHDbits <uhdbits@duck.com>