removed LineIn input
on this layout, every time when power state changed from AC to BAT, headphones were producing static noise. on previous commits i tried to fix it by setting up connector types but actually the problem was the sound card itself. unlike other laptops, this laptop puts Node 0x19 to sleep when it's not in use, to be precise when there's no sound activity. i solved this problem by removing Line In input altogether. additionally, users of this laptop must use one of the headphone patches for ALC255 (such as ALCPlugFix-Swift)